LayoutTests/fast/visual-viewport/ios/get-bounding-client-rect-unstable-expected.txt

PASS boundingClientRect.top is expectedTop
Doing unstable scroll
PASS boundingClientRect.top is expectedTop
Finishing scroll
PASS boundingClientRect.top is expectedTop
PASS successfullyParsed is true

TEST COMPLETE
